feat(booking): add unlimit, family packae into Booking context

+++ src/main/java/com/swp/autocarwash/promotion/repository/VoucherUsageRepository.java
@@ -47,5 +47,5 @@ public interface VoucherUsageRepository extends JpaRepository<VoucherUsage, Long
      * @author Phong
      * @version 1.0
      */
-    long countByVoucherIdAndCustomerId(Integer voucherId, Integer customerId);
+    long countByVoucherIdAndCustomerId(Long voucherId, Long customerId);
 }
